# docs/swagger.yaml
basePath: /
definitions:
  internal_controller.CommonError:
    description: Common error response
    properties:
      error:
        example: error message
        type: string
    type: object
  internal_controller.CommonResponse:
    description: Common success response
    properties:
      message:
        example: success message
        type: string
    type: object
  internal_controller.SetRoleRequest:
    properties:
      role:
        example: SHOPKEPER
        type: string
    required:
    - role
    type: object
...
  marketplace_internal_models_domain.Shop:
    description: Shop information
    properties:
      created_at:
        type: string
      deleted_at:
        type: string
      description:
        example: Shop description
        type: string
      id:
        example: 1
        type: integer
      name:
        example: My Shop
        type: string
      owner_id:
        example: 123
        type: integer
      slug:
        example: my-shop
        type: string
      updated_at:
        type: string
    type: object
host: localhost:7577
info:
  contact:
    email: saidnurnasreddinzoda@gmail.com
    name: API Support
  description: API Сервер для маркетплейса
  termsOfService: http://swagger.io/terms/
  title: Marketplace API
  version: "1.0"
paths:
  /api/v1/admin/users/{id}/role:
    put:
      consumes:
      - application/json
      description: Изменяет роль пользователя (только для админов)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: Новая роль
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/internal_controller.SetRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
      security:
      - BearerAuth: []
      summary: Изменить роль пользователя
      tags:
      - admin
...
  /api/v1/shops:
    get:
      description: Получает список магазинов владельца
      parameters:
      - description: Owner ID
        in: query
        name: owner_id
        required: true
        type: integer
      - default: 20
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/marketplace_internal_models_domain.Shop'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
      summary: Список магазинов
      tags:
      - shops
    post:
      consumes:
      - application/json
      description: Создает новый магазин
      parameters:
      - description: Данные магазина
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/marketplace_internal_models_domain.Shop'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/marketplace_internal_models_domain.Shop'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
      security:
      - BearerAuth: []
      summary: Создать магазин
      tags:
      - shops
  /api/v1/shops/{id}:
    delete:
      description: Удаляет магазин (soft delete, только для админов и владельцев)
      parameters:
      - description: Shop ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_controller.CommonResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
      security:
      - BearerAuth: []
      summary: Удалить магазин
      tags:
      - shops
    get:
      description: Получает информацию о магазине по его ID
      parameters:
      - description: Shop ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/marketplace_internal_models_domain.Shop'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
      summary: Получить магазин по ID
      tags:
      - shops
    put:
      consumes:
      - application/json
      description: Обновляет информацию о магазине (только для админов и владельцев)
      parameters:
      - description: Shop ID
        in: path
        name: id
        required: true
        type: integer
      - description: Данные для обновления
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/marketplace_internal_models_domain.Shop'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/marketplace_internal_models_domain.Shop'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_controller.CommonError'
      security:
      - BearerAuth: []
      summary: Обновить магазин
      tags:
      - shops
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"